refactor(stats): move the decompression profiling code into its own header
